simple-agent/requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
yfinance==0.2.18
requests==2.31.0
pytest==7.4.4
pytest-mock==3.12.0
python-dotenv==1.0.0
pydantic==2.9.2
plotly==5.17.0
streamlit==1.29.0
crewai==0.41.1
crewai-tools==0.8.3